Separate controller for CH01 and AP01 correspondence address radio button page. Update existing unit tests
