+++ src/types/page.urls.ts
@@ -198,3 +198,6 @@ export const UPDATE_DIRECTOR_SUBMITTED_PATH_END = UPDATE_PATH + "director-submit
 export const UPDATE_DIRECTOR_SUBMITTED = ACTIVE_SUBMISSION_BASE + UPDATE_DIRECTOR_SUBMITTED_PATH_END;
 export const UPDATE_DIRECTOR_SUBMITTED_PATH = OFFICER_FILING + UPDATE_DIRECTOR_SUBMITTED;
 
+export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_PATH_END = UPDATE_PATH + "director-correspondence-address"
+export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS = ACTIVE_SUBMISSION_BASE + UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_PATH_END;
+export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_PATH = OFFICER_FILING + UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS;
