+++ views/includes/check-your-answer/address-details.html
@@ -114,7 +114,7 @@
         <dd class="govuk-summary-list__value">
         </dd>
         <dd class="govuk-summary-list__actions">
-          <a class="govuk-link" id="update-director-details-link" href="{{nameLink}}" data-event-id="update-address-change-link">{{ i18n.updateDirectorDetailsUpdate }}<span class="govuk-visually-hidden">Update address</span></a>
+          <a class="govuk-link" id="update-director-details-link" href="{{correspondenceAddressChangeLink}}" data-event-id="update-address-change-link">{{ i18n.updateDirectorDetailsUpdate }}<span class="govuk-visually-hidden">Update address</span></a>
         </dd>
         <dd class="govuk-summary-list__actions">
           {% if addressHasBeenUpdated %}
