+++ views/includes/page-templates/director-correspondence-address.html
@@ -1,0 +1,93 @@
+{% extends "layout.html" %}
+
+{% set title = "What is the director's correspondence address?" %}
+
+{% block pageTitle %}
+  {% include "includes/page-title.html" %}
+{% endblock %}
+
+{% block backLink %}
+  {% set backLinkText = i18n.back %}
+  {% include "includes/back-link.html" %}
+{% endblock %}
+
+{% block content %}
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-three-quarters">
+
+      {% include "includes/list/errors.html" %}
+      {% if 'undefined' not in directorRegisteredOfficeAddress %}
+        {% set directorRegisteredOfficeAddressValue = directorRegisteredOfficeAddress %}
+      {% endif %}
+
+      
+      <form action="" method="post">
+        <span class="govuk-caption-l govuk-!-padding-bottom-1">{{directorName}}</span>
+        {{ govukRadios({
+          errorMessage: errors.director_correspondence_address if errors,
+          idPrefix: "director_correspondence_address",
+          name: "director_correspondence_address",
+          fieldset: {
+            legend: {
+              text: "What is the director's correspondence address?",
+              isPageHeading: true,
+              classes: "govuk-fieldset__legend--l"
+            }
+          },
+          hint: {
+            text: "We'll use this address to send Companies House mail to the director. You can use a business or home address and it does not have to be in the UK."
+          },
+          items: [
+            {
+              value: "director_registered_office_address",
+              text: directorRegisteredOfficeAddressValue,
+              checked: director_correspondence_address === "director_registered_office_address",
+              attributes: {
+                "data-event-id": "director_registered_office_address"
+              },
+              hint: {
+                text: "This is the company's registered office address"              
+              },
+              id: "director_registered_address"
+            },
+            {
+              value: "director_different_address",
+              text: "A different address",
+              checked: director_correspondence_address === "director_different_address",
+              attributes: {
+                "data-event-id": "director-different-correspondence-address"
+              },
+              id: "director_different_address"
+            }
+          ]
+        }) }}
+
+        <div class="govuk-inset-text">
+          <h2 class="govuk-heading-m">{{ publicRegisterInformation }}</h2>
+            <p>We will show this address on the public Companies House register.</p>
+        </div>
+
+        {% set correspondenceAddressSameAsHomeAddress %}
+        <p>We will not show your home address on the public register. However, if you use your home address as your correspondence address, we will show this on the public register. </p>
+        <p>If you run your company from home and do not want your home address available to the public, there is a  <a href="https://www.gov.uk/government/publications/formation-and-company-secretarial-agents/company-formation-agents-and-secretarial-agents" data-event-id="list-of-agents-link" class="govuk-link" target="_blank" rel="noopener noreferrer"> list of agents who can provide registered office services (opens in new tab).</a> </p> 
+        <p>A quick online search will show other agents who offer similar services. All providers will charge a fee. </p>
+        {% endset %}
+        {{ govukDetails({
+          summaryHtml: '<div> My correspondence address is the same as my home address </div>',
+          html: correspondenceAddressSameAsHomeAddress,
+          attributes: {
+            "data-event-id": "my-correspondance-address-is-my-home-address-accordion"
+          }
+        }) }}
+
+        {{ govukButton({
+          attributes: {
+            "data-event-id": "continue-button"
+          },
+          text: "Continue"
+        }) }}
+
+      </form>
+    </div>
+  </div>
+{% endblock %}
